scripts/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
psycopg2-binary==2.9.7
schedule==1.2.0
lxml==4.9.3
python-dateutil==2.8.2
